+++ frontend/src/components/host/LandingView/FloatingActionButton/index.tsx
@@ -1,26 +1,10 @@
-import { RefObject, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-
-import useOnContainerScroll from '@/hooks/useOnContainerScroll';
+import useFloatingActionButton from './useFloatingActionButton';
+import { RefObject } from 'react';
 
 import styles from './styles';
 
 const FloatingActionButton: React.FC<{ mainRef: RefObject<HTMLElement> }> = ({ mainRef }) => {
-  const navigate = useNavigate();
-  const [eventNumber, setEventNumber] = useState(0);
-
-  const { scrollInfo } = useOnContainerScroll(mainRef, () => {
-    if (scrollInfo.progress === 1) {
-      setEventNumber(1);
-      return;
-    }
-
-    setEventNumber(0);
-  });
-
-  const onClick = () => {
-    navigate('/host');
-  };
+  const { eventNumber, onClick } = useFloatingActionButton(mainRef);
 
   return (
     <div>
