+++ frontend/src/components/host/LandingView/HostViewSection1/index.tsx
@@ -1,25 +1,11 @@
-import { useRef, useState } from 'react';
-
-import useOnContainerScroll from '@/hooks/useOnContainerScroll';
+import useHostViewSection1 from './useHostViewSection1';
 
 import { ScreenModeType } from '@/types';
 
 import styles from './styles';
 
 const HostViewSection1 = ({ screenMode }: { screenMode: ScreenModeType }) => {
-  const [eventNumber, setEventNumber] = useState(0);
-  const sectionRef = useRef<HTMLDivElement>(null);
-
-  const { dimension, scrollInfo } = useOnContainerScroll(sectionRef, () => {
-    const progress = Math.min(Math.max(0, scrollInfo.scrollY / dimension.windowHeight), 1) * 100;
-    if (progress === 0) return;
-
-    if (progress >= 80) {
-      setEventNumber(1);
-      return;
-    }
-    setEventNumber(0);
-  });
+  const { eventNumber, sectionRef } = useHostViewSection1();
 
   return (
     <section css={styles.layout(screenMode)} ref={sectionRef}>
