+++ frontend/src/components/host/LandingView/HostViewSection3/index.tsx
@@ -1,26 +1,11 @@
-import { useRef, useState } from 'react';
-
-import useOnContainerScroll from '@/hooks/useOnContainerScroll';
+import useHostViewSection3 from './useHostViewSection3';
 
 import { ScreenModeType } from '@/types';
 
 import styles from './styles';
 
 const HostViewSection3 = ({ screenMode }: { screenMode: ScreenModeType }) => {
-  const [eventNumber, setEventNumber] = useState(0);
-  const sectionRef = useRef<HTMLDivElement>(null);
-
-  const { dimension, scrollInfo } = useOnContainerScroll(sectionRef, () => {
-    const progress = Math.min(Math.max(0, scrollInfo.scrollY / dimension.windowHeight), 1) * 100;
-    if (progress === 0) return;
-
-    if (progress >= 20) {
-      setEventNumber(1);
-      return;
-    }
-
-    setEventNumber(0);
-  });
+  const { eventNumber, sectionRef } = useHostViewSection3();
 
   return (
     <section css={styles.layout} ref={sectionRef}>
