+++ frontend/src/components/host/LandingView/UserViewSection1/index.tsx
@@ -1,7 +1,5 @@
 import UserPicture from '../UserPicture';
-import { useRef, useState } from 'react';
-
-import useOnContainerScroll from '@/hooks/useOnContainerScroll';
+import useUserViewSection1 from './useUserViewSection1';
 
 import { ScreenModeType, UserImageType } from '@/types';
 
@@ -35,23 +33,7 @@ const mobileView2: UserImageType = {
 };
 
 const UserViewSection1 = ({ screenMode }: { screenMode: ScreenModeType }) => {
-  const [eventNumber, setEventNumber] = useState(0);
-  const sectionRef = useRef<HTMLDivElement>(null);
-
-  const { dimension, scrollInfo } = useOnContainerScroll(sectionRef, () => {
-    const progress = Math.min(Math.max(0, scrollInfo.scrollY / dimension.windowHeight), 1) * 100;
-    if (progress === 0) return;
-
-    if (progress >= 40 && progress < 80) {
-      setEventNumber(1);
-      return;
-    }
-    if (progress >= 80) {
-      setEventNumber(2);
-      return;
-    }
-    setEventNumber(0);
-  });
+  const { eventNumber, sectionRef } = useUserViewSection1();
 
   return (
     <section css={styles.layout} ref={sectionRef}>
